server/requirements.txt
# Core Application
fastapi==0.111.1
uvicorn==0.30.3
websockets==12.0
pydantic==2.8.2
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data & Scientific Computing
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1
torch==2.5.0
cupy-cuda12x==13.2.0

# Celery & Background Jobs
celery==5.4.0
redis==5.0.7

# Performance & Load Testing
locust==2.29.0
gevent==25.5.1
psutil==5.9.8

# Data Export & Plotting
openpyxl==3.1.5
plotly==5.22.0

# Development & Testing
pytest==8.4.1
requests==2.32.4 